added product fetching, moved old product fetch to unused folder

+++ pages/api/products/index.ts
@@ -6,36 +6,40 @@ export default async function handler(
   request: NextApiRequest,
   response: NextApiResponse
 ) {
-  try {
-    const client = await clientPromise;
-    const db = client.db("shop");
+  const { method } = request;
+  const client = await clientPromise;
+  const db = client.db("shop");
 
-    // PUT: uses req.body.param.id
-    // DELETE: uses req.query.id
-
-    const { data } = request.body;
-
-    switch (request.method) {
-      case "GET":
-        const result = await db.collection("products").find({}).toArray();
-        const products = JSON.parse(JSON.stringify(result));
-        return response.status(200).json(products);
-
-      case "POST":
-        await db.collection("products").insertOne(data);
-        return response.status(201).json({ message: "product created" });
-
-      case "DELETE":
-        await db
-          .collection("products")
-          .deleteOne({ _id: new ObjectId(request.query.id as string) });
-
-        return response.status(200).json({ message: "product deleted" });
-
-      default:
-        break;
+  if (method === "GET") {
+    try {
+      const result = await db.collection("products").find({}).toArray();
+      const products = JSON.parse(JSON.stringify(result));
+      return response.status(200).json(products);
+    } catch (error) {
+      response.status(500).json(error);
     }
-  } catch (error) {
-    response.status(500).json({ message: "something went wrong", error });
+  }
+
+  if (method === "POST") {
+    try {
+      const { data } = request.body;
+      // console.log(request.body) // use request.body with postman
+      await db.collection("products").insertOne(data);
+      return response.status(201).json({ message: "Product created." });
+    } catch (error) {
+      response.status(500).json(error);
+    }
+  }
+
+  if (method === "DELETE") {
+    try {
+      await db
+        .collection("products")
+        .deleteOne({ _id: new ObjectId(request.query.id as string) });
+
+      return response.status(200).json({ message: "Product deleted." });
+    } catch (error) {
+      response.status(500).json(error);
+    }
   }
 }
